Split handler errors
